Several bugfixes related to CSR storage format (not completed yet ...)
The currently available CSR implementation seems to target only a pre_to_post matrix order, which is obviously problematic. I particularly fixed that, but the CSR implementation is still in an unstable state ...
I also disabled the proj._dense mode as it is for LIL format (default) completely without effect and leads to errors in case of CSR format.